requirements.txt
textual
numpy
keyboard
mouse
clipboard
pyautogui
pillow
qrcode
DynamicStateMachine>=0.0.3
streamlit-keyup
